Move prisma client file
