+++ lib/serverAuth.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest } from "next";
 import { getSession } from "next-auth/react";
-import prismadb from "@/lib/prismadb";
+import prismadb from "@/lib/prisma";
 
 const serverAuth = async (req: NextApiRequest) => {
   const session = await getSession({ req });
