+++ pages/api/auth/[...nextauth].ts
@@ -1,6 +1,6 @@
 import NextAuth from "next-auth/next";
 import Credentials from "next-auth/providers/credentials";
-import prismadb from "@/lib/prismadb";
+import prismadb from "@/lib/prisma";
 import { compare } from "bcrypt";
 import GithubProvider from "next-auth/providers/github";
 import GoogleProvider from "next-auth/providers/google";
